+++ benchmark/kd_tree/kd_tree_building.dart
@@ -1,10 +1,8 @@
-// 0.8 sec (MacBook Air mid 2017)
-import 'dart:convert';
-import 'dart:io';
-
+// 0.5 sec (MacBook Air mid 2017)
 import 'package:benchmark_harness/benchmark_harness.dart';
 import 'package:ml_algo/src/retrieval/kd_tree/kd_tree.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
+import 'package:ml_linalg/matrix.dart';
 
 late DataFrame trainData;
 
@@ -24,11 +22,9 @@ class KDTreeBuildingBenchmark extends BenchmarkBase {
 }
 
 Future main() async {
-  final file = File('benchmark/data/sample_data.json');
-  final dataAsString = await file.readAsString();
-  final decoded = jsonDecode(dataAsString) as Map<String, dynamic>;
+  final points = Matrix.random(1000, 10, seed: 1, min: -5000, max: 5000);
 
-  trainData = DataFrame.fromJson(decoded);
+  trainData = DataFrame.fromMatrix(points);
 
   print(
       'Data dimension: ${trainData.rows.length}x${trainData.rows.first.length}');