+++ lib/src/retrieval/kd_tree/helpers/create_kd_tree_from_iterable.dart
@@ -2,12 +2,16 @@ import 'package:ml_algo/src/retrieval/kd_tree/kd_tree_builder.dart';
 import 'package:ml_algo/src/retrieval/kd_tree/kd_tree_constants.dart';
 import 'package:ml_algo/src/retrieval/kd_tree/kd_tree_impl.dart';
 import 'package:ml_algo/src/retrieval/kd_tree/kd_tree_split_strategy.dart';
-import 'package:ml_dataframe/ml_dataframe.dart';
 import 'package:ml_linalg/dtype.dart';
+import 'package:ml_linalg/matrix.dart';
 
-KDTreeImpl createKDTree(DataFrame pointsSrc, int leafSize, DType dtype,
-    KDTreeSplitStrategy splitStrategy) {
-  final points = pointsSrc.toMatrix(dtype);
+KDTreeImpl createKDTreeFromIterable(Iterable<Iterable<num>> pointsSrc,
+    int leafSize, DType dtype, KDTreeSplitStrategy splitStrategy) {
+  final points = Matrix.fromList(
+      pointsSrc
+          .map((row) => row.map((element) => element.toDouble()).toList())
+          .toList(),
+      dtype: dtype);
   final builder = KDTreeBuilder(leafSize, points, splitStrategy);
   final root = builder.train();