+++ test/retrieval/kd_tree/kd_tree_test.dart
@@ -1,6 +1,7 @@
 import 'package:ml_algo/src/retrieval/kd_tree/exceptions/invalid_query_point_length.dart';
 import 'package:ml_algo/src/retrieval/kd_tree/kd_tree.dart';
 import 'package:ml_algo/src/retrieval/kd_tree/kd_tree_impl.dart';
+import 'package:ml_algo/src/retrieval/kd_tree/kd_tree_split_strategy.dart';
 import 'package:ml_dataframe/ml_dataframe.dart';
 import 'package:ml_linalg/dtype.dart';
 import 'package:ml_linalg/vector.dart';
@@ -36,7 +37,7 @@ void main() {
     test(
         'should find the closest neighbours for [2.79, -9.15, 6.56, -18.59, 13.53], leafSize=3',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       final sample = Vector.fromList([2.79, -9.15, 6.56, -18.59, 13.53]);
       final result = kdTree.query(sample, 3).toList();
 
@@ -49,7 +50,21 @@ void main() {
     test(
         'should find the closest neighbours for [2.79, -9.15, 6.56, -18.59, 13.53], leafSize=1',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 1);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 1);
+      final sample = Vector.fromList([2.79, -9.15, 6.56, -18.59, 13.53]);
+      final result = kdTree.query(sample, 3).toList();
+
+      expect(result[0].index, 12);
+      expect(result[1].index, 4);
+      expect(result[2].index, 18);
+      expect(result, hasLength(3));
+    });
+
+    test(
+        'should find the closest neighbours for [2.79, -9.15, 6.56, -18.59, 13.53], leafSize=3, splitStrategy=KDTreeSplitStrategy.inOrder',
+        () {
+      final kdTree = KDTree(DataFrame(data, headerExists: false),
+          leafSize: 3, splitStrategy: KDTreeSplitStrategy.inOrder);
       final sample = Vector.fromList([2.79, -9.15, 6.56, -18.59, 13.53]);
       final result = kdTree.query(sample, 3).toList();
 
@@ -62,7 +77,7 @@ void main() {
     test(
         'should find the closest neighbours for [13.98, -8.21, 17.01, -5.14, 14.49], leafSize=3',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       final sample = Vector.fromList([13.98, -8.21, 17.01, -5.14, 14.49]);
       final result = kdTree.query(sample, 4).toList();
 
@@ -76,7 +91,7 @@ void main() {
     test(
         'should find the closest neighbours for [13.98, -8.21, 17.01, -5.14, 14.49], leafSize=1',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 1);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 1);
       final sample = Vector.fromList([13.98, -8.21, 17.01, -5.14, 14.49]);
       final result = kdTree.query(sample, 4).toList();
 
@@ -90,7 +105,7 @@ void main() {
     test(
         'should find the closest neighbours for [-9.88, -5.66, -16.15, 4.46, 2.34]',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       final sample = Vector.fromList([-9.88, -5.66, -16.15, 4.46, 2.34]);
       final result = kdTree.query(sample, 20).toList();
 
@@ -103,7 +118,7 @@ void main() {
     });
 
     test('should find the closest neighbours for [1, 2, 3, 4, 5]', () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       final sample = Vector.fromList([1, 2, 3, 4, 5]);
       final result = kdTree.query(sample, 1).toList();
 
@@ -112,7 +127,7 @@ void main() {
     });
 
     test('should find the closest neighbours for [100, -222, 444, 0, 1]', () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       final sample = Vector.fromList([100, -222, 444, 0, 1]);
       final result = kdTree.query(sample, 1).toList();
 
@@ -123,7 +138,7 @@ void main() {
     test(
         'should find the closest neighbours for [7.63, -10.70, 15.09, 10.25, -18.16] for concievable amount of iterations, leafSize=1',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 1);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 1);
       final sample = Vector.fromList([7.63, -10.70, 15.09, 10.25, -18.16]);
 
       kdTree.query(sample, 1).toList();
@@ -134,7 +149,7 @@ void main() {
     test(
         'should find the closest neighbours for [7.63, -10.70, 15.09, 10.25, -18.16] for concievable amount of iterations, leafSize=3',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       final sample = Vector.fromList([7.63, -10.70, 15.09, 10.25, -18.16]);
 
       kdTree.query(sample, 1).toList();
@@ -144,7 +159,7 @@ void main() {
 
     test('should throw an exception if the query point is of invalid length',
         () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       final sample = Vector.fromList([1, 2, 3, 4, 5, 6]);
 
       expect(() => kdTree.query(sample, 1),
@@ -152,7 +167,7 @@ void main() {
     });
 
     test('should persist points', () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
 
       expect(
           kdTree.points,
@@ -181,13 +196,45 @@ void main() {
     });
 
     test('should persist dtype', () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       expect(kdTree.dtype, DType.float32);
     });
 
     test('should persist leaf size', () {
-      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSie: 3);
+      final kdTree = KDTree(DataFrame(data, headerExists: false), leafSize: 3);
       expect(kdTree.leafSize, 3);
+    });
+
+    test('should create from iterable', () {
+      final kdTree =
+          KDTree.fromIterable(data, dtype: DType.float64, leafSize: 13);
+
+      expect(
+          kdTree.points,
+          iterable2dAlmostEqualTo([
+            [3.43, 10.91, 11.62, -12.93, -11.66],
+            [19.41, -4.96, 3.99, 16.35, 10.57],
+            [11.30, 8.89, -17.66, -5.17, 16.20],
+            [-8.13, -5.23, 18.01, 1.97, 9.08],
+            [13.98, -8.21, 17.01, -5.14, 14.49],
+            [-17.65, 13.10, 5.82, 8.61, 14.41],
+            [4.16, -4.72, -3.71, -2.32, -13.70],
+            [7.29, 11.16, -9.51, -1.89, -18.94],
+            [19.81, 3.17, 14.27, 0.05, -17.93],
+            [-9.63, 18.82, -14.40, -1.91, -6.58],
+            [-10.95, -19.58, 9.05, 17.39, 3.30],
+            [4.08, -13.19, -5.71, 18.56, -0.13],
+            [2.79, -9.15, 6.56, -18.59, 13.53],
+            [-7.56, 11.97, 6.55, -7.54, 15.90],
+            [-15.97, -15.95, 7.71, 9.70, 16.94],
+            [-15.01, 16.12, -10.42, -17.61, 6.27],
+            [7.63, -10.70, 15.09, 10.25, -18.16],
+            [0.05, 9.74, 7.08, 15.49, -17.99],
+            [-6.48, 1.10, 9.28, 0.90, 6.09],
+            [-9.88, -5.66, -16.15, 4.46, 2.34],
+          ]));
+      expect(kdTree.dtype, DType.float64);
+      expect(kdTree.leafSize, 13);
     });
   });
 }